fix(parity): Thought for 1.2s + Turn completed in 3.6s now render in replay
Two real bugs fixed:

1. **Thought duration skipped in replay.** In scenario_replay/tests,
   `ThinkingEnd` arrives BEFORE the matching `MessageStart` (no
   submit() happens). The handler was looking for an assistant
   message that didn't exist yet, so the duration got dropped.
   Added `state.pending_thought_duration` — on_thinking_end
   stashes the duration there, on_message_start consumes it
   when the assistant placeholder is created.

2. **Turn completed in 0.0s in replay.** `on_turn_end` only ran
   the separator push if `state.agent_start_time` was set (a
   user submit). In replay there's no submit, so nothing got
   pushed. Added `state.replay_turn_duration_secs` (f32) that
   on_thinking_end writes and on_turn_end reads, preserving
   sub-second precision (3.6s, 1.2s) instead of truncating to
   integer seconds.

   Also: `last_turn_duration_secs` was being clobbered to None
   by a second `ThinkingStart` event in the same turn. The new
   `replay_turn_duration_secs` field is only set, never reset.

3. **Bonus: turn_complete (f32) preserved end-to-end.**
   The renderer was casting f32 turn_duration → u64 (truncating
   3.6s → 3s). Changed render_assistant_msg's `turn_complete`
   param from `Option<u64>` to `Option<f32>` so the "{:.1}s"
   formatter shows 3.6, not 3.0.

Scenario updated to use 3600ms thinking to match the Grok
reference dump (was 1200ms).

+++ crates/runie-tui/src/components/message_list/render/assistant.rs
@@ -139,7 +139,7 @@ pub fn render_assistant_msg(
     spinner: char,
     timestamp: Option<&str>,
     thought_duration: Option<f32>,
-    turn_complete: Option<u64>,
+    turn_complete: Option<f32>,
     _is_last_item: bool,
     tool_calls: &[ToolCall],
     tool_bar_color: ratatui::style::Color,
